Eigensolvers

Recovering the roots from operators of multiplication is done by join diagonalization when the roots are simple.

AlgebraicSolvers.diagonalization — Function

Xi, E, Info =  diagonalization(M::Vector{Matrix{C}}; maxiter::Int64 = 10, epsiter::Float64 = 1.e-3)

Compute the joint diagonalizaion of the matrices in M (assuming it exists). maxiter is the maximum number of iterations and epsiter is the threshold error to stop the iterations. It outputs

- `Xi` the diagonals row by row, the ith row for the matrix  `M[i]`
- `E`  the matrix of common eigenvectors so that `M[i] = E*diagm(Xi[i,:])*inv(E)`

AlgebraicSolvers.eigdiag — Function

Xi, E =  eigdiag(M::Vector{Matrix{C}})

Compute the joint diagonalizaion of the matrices in M, from the computation of the eigen vectors of a random combination of the matrices in M.It outputs

- `Xi` the points in diagonals of the `E^(-1)*M*E`
- `E`  the matrix of  eigenvectors of a random combination of `M`, so that `M[i] = E*diagm(Xi[i,:])*inv(E)`

When the roots are multiple, one can use the joint Schur factorization:

AlgebraicSolvers.schur_dcp — Function

Xi, ms, Z =  schur_dcp(M::AbstractVector, eps::Float64=1.e-5)

Compute the joint Schur factorization of M from the Schur factorization of a random combination of the M[i]. The eigenvalues are clustered within the threshold eps, using the function multiplicities. It outputs

- `Xi` the points from trace of the Schur blocks
- `ms` the indices of the blocks corresponding to the points `Xi`
- `Z`  the Schur factor

The clustering of eigenvalues is done with the following function:

AlgebraicSolvers.multiplicities — Function

ms = multiplicities(v::AbstractVector, eps = 1.e-5)

Group the values in v by clusters of radius < eps and output the list of lists of indices of the clusters

Example

julia> multiplicities([0.,3.,0.0000000001]) 2-element Vector{Vector{Int64}}: [1, 3] [2]