Solvers

Several types of solvers are available in the packages. They differ in the way the truncated normal form is computed.

AlgebraicSolvers.solve — Function

solve(Mth, P; verbose = false)

P polynomial system

Solve the system or polynomials P=[p1, ..., pn], building Sylvester matrix res_matrix(Mth,P).

Example

using AlgebraicSolvers, DynamicPolynomials

X = @polyvar x y

P = [y - x*y + x^2+ y^2,  1 + y + x + x^2]

Xi = solve(Macaulay(), P; verbose=true)

Xi = solve(Macaulay(), P)

P polynomial system

Solve the system P=[p1, ..., pn], building Sylvester matrix of all monomial multiples mi*pi in degree ≤ ρ.

The default value for ρ is ∑ deg(pi) - n + 1.

Example

using AlgebraicSolvers, DynamicPolynomials

X = @polyvar x y

P = [2-x*y+x^2,y^2+x-2]

Xi = solve(Macaulay(), P)

Xi, G, B = solve(Mth::Grobner, P::Vector; verbose = false)

Solve the system of polynomials P. It outputs:

Xi the complex solution points, one per column of Xi

G the computed Grobner basis

B the basis of the quotient by the ideal of the equations

If verbose = true, the timing of the different steps is printed.

Example:

using AlgebraicSolvers, DynamicPolynomials, Groebner

Mth = Grobner((P,X) -> Groebner.groebner(P, ordering = Groebner.DegRevLex(X)),
              (p,G) -> Groebner.normalform(p,G),
              G -> Groebner.quotient_basis(G)
             )

X = @polyvar x y 

P = [-2+y-y^2+x^2*y, 1-3x+y+x*y^2]

Xi, G, B = AlgebraicSolvers.solve(Mth, P; verbose = true)

using AbstractAlgebra

R ,(x,y) =  polynomial_ring(AbstractAlgebra.QQ, [:x,:y])

P = [x^2*y-y^2, x*y^2-3]

Xi, G, B = AlgebraicSolvers.solve(Mth, P)

Solvers using resultant constructions

They corresponds to the following type

AlgebraicSolvers.Macaulay — Type

Structure for the construction of Macaulay resultant solvers. It stores

degree :: Function P-> degree of regularity +1 (default: P -> sum(DynamicPolynomials.maxdegree(P[i])-1 for i in 1:length(P)) + 1)

is_homogeneous :: Function P-> boolean testing if the system is homogeneous or not (default: P -> !any(AlgebraicSolvers.is_not_homogeneous, P))

AlgebraicSolvers.Toric — Type

Structure for the construction of Toric resultant solvers. It store

supports :: Function P-> supports of the polynomials P (default: P->AlgebraicSolvers.support.(P) )

Solvers using Groebner basis computation

AlgebraicSolvers.Grobner — Type

Structure which defines

grobner_basis :: Function (P,X) -> grobner basis of P with variables X

reduce :: Function (p, G) -> normal form of the polynomial p modulo G

quotient_basis :: Function P -> basis of the quotient by the ideal(P)

Quotient Algebra

Truncated Normal Forms, which kernel generates the ideal, are available:

AlgebraicSolvers.tnf — Function

N, L = tnf(Mth, P)

Compute the Truncated Normal Form N of P=[p1, ..., pn], using res_matrix(Mth,P).

The list L is the list of monomials indexing the colmuns of N.

    B = tnf(Mth::Grobner,P)

Computes the Truncated Normal Form on B^+ where B is the basis of the quotient by the ideal (P), formed by the monomials which are not in the inital of (P).

The operators of multiplications (by the variables or any polynomial) can be computed with the following function:

AlgebraicSolvers.mult_matrices — Function

    M = mult_matrices(Mth, P, A = DynamicPolynomials.variables(P))

Computes the vector of multiplication matrices M=[1, M2, ...] by the variables in a basis B of the quotient by the ideal (P), using res_matrix(Mth,P).

AlgebraicSolvers.mult_matrix — Function

M = mult_matrix(M::Grobner, p, G::AbstractVector, Idx::Dict)

Compute the matrix of multication by p modulo g in the basis associated to the basis dictionary Idx. It is assumed that g is a Groebner basis and that the quotient is finite dimensional.

M = mult_matrix(p, G::AbstractVector, B::AbstractVector)

Compute the matrix of multication by p modulo G in the basis B. It is assumed that G is a Groebner basis and that the quotient is finite dimensional.

A basis of the quotient by the ideal (P) can be also computed:

AlgebraicSolvers.quo_basis — Function

    B = quo_basis(Mth, P)

Computes a basis of the quotient by the ideal P, using res_matrix(Mth,P)

    B = quo_basis(Mth::Grobner,P)

Computes the basis B of the quotient by the ideal (P), formed by the monomials which are not in the inital of (P).