Eigen solvers

AlgebraicSolvers.diagonalization — Function

Xi, E, Info =  diagonalization(M::Vector{Matrix{C}}; maxiter::Int64 = 10, epsiter::Float64 = 1.e-3)

Compute the joint diagonalizaion of the matrices in M (assuming it exists). maxiter is the maximum number of iterations and epsiter is the threshold error to stop the iterations. It outputs

- `Xi` the diagonals row by row, the ith row for the matrix  `M[i]`
- `E`  the matrix of common eigenvectors so that `M[i] = E*diagm(Xi[i,:])*inv(E)`