Series

Series are sequences of values indexed by monomials.

\[ \sigma = \sum_{\alpha \in A \subset \mathbb{N}^n } \sigma_{\alpha} \mathbf{y}_1^{\alpha_1} \cdots \mathbf{y}_n^{\alpha_n}\]

They represent linear functionals on polynomials, or dual elements of the polynomial ring:

\[ \sigma: p = \sum_{\alpha} p_{\alpha} \mathbf{x}^{\alpha} \mapsto \langle \sigma , p \rangle = \sum_{\alpha} \sigma_{\alpha}\, p_{\alpha}\]

They are implemented as dictionaries between monomials and coefficients

AlgebraicSolvers.Series — Type

Series{C,M}

Class representing multivariate series. The series is a dictionary, which associates values of type C to monomials of type M.

AlgebraicSolvers.series — Function

series(c::C, m::M)

Construct the series with the term (c,m).

series(t::OrderedDict{M,C})

Construct the series from the OrderedDict t.

series(c::AbstractVector, m::AbstractVector)

Construct the series with the vectors of monomials m and coefficients c.

series(s::Vector{Pair{M,C}})

Construct the series from an array of pairs m=>c where m is a monomial and c the associate coefficient.

series(f::Function, L::Vector{M}) -> Series{C,M}

Compute the generating series $\sum_{x^{α} \in L} f(α) z^α$ for a function $f: \mathbb{N}^n \rightarrow C$ and a sequence L of monomials.

series(w:: Vector{C}, P::Matrix{C}, L::Vector{M}) -> Series{C,M}

Compute the series of the moment sequence $∑_{i} ω_{i} P_{i}^α$ for $α \in L$.

series(w::AbstractVector, P::AbstractMatrix, X, d::Int64) -> Series{C,M}

Compute the series of the moment sequence $∑_i ω_{i} P_{i}^α$ for $|α| \leq d$.

series(p::Polynomial, zeta, X, d::Int64) -> Series

Compute the series of moments $p(ζ^α)$ for $|α| \leq d$.

series(H::Matrix{C}, L1::Vector{M}, L2::Vector{M}) -> Series{C,M}

Compute the series associated to the Hankel matrix H, with rows (resp. columns) indexed by the array of monomials L1 (resp. L2).

AlgebraicSolvers.dual — Function

s = dual(p::Polynomial) -> Series{C,M}

Compute the series sassociated to the polynomial p, replacing the variables xi by their dual variables dxi. C is the type of coefficients of the polynomial p and M its type of monomials.

Recover the polynomial from the series by duality.

Recover the polynomial from the series, using the apolar duality i.e. multiplying the coefficients by a binomial.

AlgebraicSolvers.apolar_dual — Function

    s = apolar_dual(p::AbstractPolynomial)

Compute the apolar series s associated to the symmetric tensor p in degree d. The type of the coefficients of s is the promoted type between the type of the coefficients of p and Rational{Int}. The coefficients of the monomials m of p are divided by the binomial(d, exponents(m)).

    s = apolar_dual(F::AbstractPolynomial, X0; rescaling = 1)

Compute the series s from the apolar dual of the form F in degree d, setting X0=>1 and rescalling the other variables X[i] => X[i]/rescaling. The type of the coefficients of s is the promoted type between the type of the coefficients of p and Rational{Int}.

AlgebraicSolvers.moment — Function

moment(w::Vector{C}, P::Matrix{C}) -> Vector{Int64} -> C

Compute the moment function $α -> ∑_{i} ω_{i} P_{i}^α$ associated to the sequence P of r points of dimension n, which is a matrix of size r*n and the weights w.

moment(p::Polynomial, zeta::Vector{C}) -> Vector{Int64} -> C

Compute the moment function $α \rightarrow p(ζ^α)$.

The apolarity involves the multi-index binomial:

Base.binomial — Function

Base.binomial(d::Int64, alpha::Vector{Int64})

Multi-index binomial coefficients: $\frac{d!}{\prod_i \alpha_i!}$. One should have $\alpha_i \ge 0$and $\sum_i \alpha_i \le d$.

AlgebraicSolvers.scale — Function

scale(σ,λ)

Scale the moments $σ_α$ by $λ^{deg(α)}$.

AlgebraicSolvers.scale! — Function

scale!(σ,λ)

Scale the moments $σ_α$ by $λ^{deg(α)}$, overwriting $σ$

Base.:* — Function

Multiply the elements of L by the variable v

     *(v::Variable,   σ::Series{C,M}) -> Series{C,M}
     *(m::Monomial,   σ::Series{C,M}) -> Series{C,M}
     *(t::Term,       σ::Series{C,M}) -> Series{C,M}
     *(p::Polynomial, σ::Series{C,M}) -> Series{C,M}

The dual product (or co-product) where variables are inverted in the polynomial and the monomials with positive exponents are kept in the series.

Base.:& — Function

    s & p

Term-wise product of series s by p. All the terms of s are multiplied by p If σ = ∑ σm m, s & p = ∑m (σ_m*p) m

LinearAlgebra.dot — Function

dot(σ::Series{C,M}, p::Variable) -> C
dot(σ::Series{C,M}, p::Monomial) -> C
dot(σ::Series{C,M}, p::Term) -> C
dot(σ::Series{C,M}, p::Polynomial) -> C
dot(σ::Series{C,M}, p::Polynomial, q::Polynomial) -> C

Compute the dot product $‹ p, q ›_{σ} = ‹ σ | p q ›$ or $‹ σ | p ›$ for p, q polynomials, terms or monomials. Apply the linear functional sigma on monomials, terms, polynomials

Base.:| — Function

Apply the linear functional sigma to p.

AlgebraicSolvers.hankel — Function

hankel(σ::Series{C,M}, L1::Vector{M}, L2::Vector{M}) -> Array{C,2}

Hankel matrix of $σ$ with the rows indexed by the list of polynomials L1 and the columns by L2. The entries are the dot product for $σ$ of the corresponding elements in L1 and L2.

Example

julia> L =[1, x1, x2, x1^2, x1*x2, x2^2]

julia> H = hankel(s,L,L)
6x6 Array{Float64,2}:
  4.0   5.0   7.0    5.0  11.0  13.0
  5.0   5.0  11.0   -1.0  17.0  23.0
  7.0  11.0  13.0   17.0  23.0  25.0
  5.0  -1.0  17.0  -31.0  23.0  41.0
 11.0  17.0  23.0   23.0  41.0  47.0
 13.0  23.0  25.0   41.0  47.0  49.0

    hankel(p::AbstractPolynomial, d1:: Int64, d2::Int64)

Compute the so-called Catalecticant matrix of the homogeneous polynomial p in bidegree (d1, d2). One should have maxdegree(p)=d1+d2.

AlgebraicSolvers.hankelbasis — Function

Generate the table of Hankel matrices $H_α$ associated to the monomials $x^α$ generating the space of Hankel matrices indexed by L1 (for the rows) and L2 (for the columns).

MultivariatePolynomials.maxdegree — Function

    maxdegree(σ::Series) -> Int64

Maximal degree of the monomials of the moments of the series σ.

AlgebraicSolvers.prodset — Function

prodsect(X,Y)

Product-wise vector of X by Y ordered by row with no repetition: if xiyj appears as xi'yj' with i'<i or i'=i and j'<j it is not inserted.

AlgebraicSolvers.prodvec — Function

prodvec(X,Y)

Product-wise vector of X by Y ordered by row: [x1y1, x1y2, ..., x2*y1, ....]

AlgebraicSolvers.integrate — Function

Compute the primitive $\int_v s = v^{-1}*s$

Compute the truncated primitive ``\int{x{i}} s{| x{i+1}=0, ...x_n=0}''

Base.inv — Function

    inv(m :: Monomial)

return the inverse monomial with opposite exponents.

 inv(m :: Monomial{true})

return the inverse monomial with opposite exponents.