Functions

AlgebraicSolvers.rel_error — Function

Vector of relative errors of P at the points X

AlgebraicSolvers.alpha_beta — Function

alpha, beta quantities for Newton convergence to an approximate zero.

If alpha < 0.125, then the approximate zero is withing 2*beta from Xi and Newton methods converges to it from Xi quadratically.

If alpha < 0.02, then Newton method converges from all points in the ball of center Xi and radius 2*beta.