Solvers

Several types of solvers are available in the packages. They differ in the way the truncated normal form is computed.

Solvers using resultant constructions

AlgebraicSolvers.solve_macaulay — Function

Xi = solve_macaulay(P, rho ; verbose = false)

P polynomial system

rho (optional) degree of regularity for the Sylvester matrix construction (optional)

Solve the system P=[p1, ..., pn], building Sylvester matrix of all monomial multiples mi*pi in degree ≤ ρ.

The default value for ρ is ∑ deg(pi) - n + 1.

Example

using AlgebraicSolvers, DynamicPolynomials

X = @polyvar x y

P = [2-x*y+x^2,y^2+x-2]

Xi = solve_macaulay(P)

AlgebraicSolvers.solve_toric — Function

solve_toric(P; verbose = false)

P polynomial system

X array of variables

Solve the system P=[p1, ..., pn], building Sylvester matrix of all monomial multiples mi*pi for mi in supp(∏_{j != i} pj).

Example

using AlgebraicSolvers, DynamicPolynomials

X = @polyvar x y

P = [y - x*y + x^2,  1 + y + x + x^2]

Xi = solve_toric(P)

Solvers using Groebner basis computation

AlgebraicSolvers.solve_groebner — Function

Xi, G, B = solve_groebner(P::Vector; verbose = false)

Solve the system of polynomials P. It outputs:

Xi the complex solution points, one per column of Xi

G the computed Groebner basis

B the basis of the quotient by the ideal of the equations

Example:

using AbstractAlgebra

R, (x,y) = QQ["x","y"]

P = [x^2-y, x^2*y-4]

Xi, G, B = solve_groebner(P)

Resultant matrices

Different constructions of resultant matrices are available, including projective or Macaulay resultant matrices, toric resultant matrices.

AlgebraicSolvers.matrix_macaulay — Function

R, L = matrix_macaulay(P, X, rho, ish = false)

P polynomial system

X (optional) array of variables

rho (optional) maximal degree of the multiples of P

ish (optional) set to true if the polynomials are homogeneous

It outputs

R the transpose of Sylvester matrix of all monomial multiples mi*pi in degree ≤ rho.

L array of monomials indexing the columns of R

AlgebraicSolvers.matrix_toric — Function

R, L = matrix_toric(P, A)

where

P polynomial system

A array of supports of Pi

It outputs

R transpose of the Sylvester matrix of all monomial multiples mi*pi for mi in supp(∏_{j != i} pj).

L the list of monomials indexing the colums of R

Truncated Normal Forms

AlgebraicSolvers.tnf_macaulay — Function

N, L = tnf_macaulay(P, rho)

Compute the Truncated Normal Form of P=[p1, ..., pn], using Macaulay matrix of all monomial multiples mi*pi in degree ≤ ρ.

The default value for ρ is ∑ deg(pi) - n + 1.

AlgebraicSolvers.tnf_toric — Function

N, L = tnf_toric(P, A = support.(P))

Compute the Truncated Normal Form of P=[p1, ..., pn], using toric resultant matrix of all monomial multiples mi*pi in degree ≤ ρ.

The default value for ρ is ∑ deg(pi) - n + 1.

Linear algebra in quotient

AlgebraicSolvers.matrix_mult — Function

M = matrix_mult(p, g::AbstractVector, Idx::Dict)

Compute the matrix of multication by p modulo g in the basis associated to the basis dictionary Idx. It is assumed that g is a Groebner basis and that the quotient is finite dimensional.

M = matrix_mult(p, G::AbstractVector, B::AbstractVector)

Compute the matrix of multication by p modulo G in the basis B. It is assumed that G is a Groebner basis and that the quotient is finite dimensional.

AlgebraicSolvers.quotient_basis — Function

B, Idx = quotient_basis(g::AbstractVector)

Compute the monomial basis of the quotient by g, as the complementary of the initial ideal of g, assuming that g is a Groebner basis and it is finite. It outputs:

B the basis of monomlials in increasing monomial order.

Idx the dictionary associting to a monomial its index in the basis.