Inverse systems

Inverse systems can be computed to analyze the multiplicity structure of multiple roots.

AlgebraicSolvers.invsys — Function

D,B,h,Nu = invsys(F::Vector, Xi::Vector)

Compute the inverse system of the polynomial system F at the point Xi. It outputs

D a basis of the inverse system as series in the dual basis of the monomial basis.

B the dual monomial basis

h the dimensions of the spaces in each degree

Nu the indices of the unknowns at each integration

using DynamicPolynomials, AlgebraicSolvers

X = @polyvar x1 x2

F = [x1 - x2 + x1^2*1.0, x1 - x2 + x2^2*1.0]

D, B, h, Nu = invsys(F,[0,0];verbose=true)
D

3-element Vector{Series{Float64, DynamicPolynomials.Monomial{DynamicPolynomials.Commutative{DynamicPolynomials.CreationOrder}, Graded{LexOrder}}}}:
 1.0
 dx1 + dx2
 dx1^2 + dx2 + dx1*dx2 + dx2^2