Solvers

Several types of solvers are available in the packages. They differ in the way the truncated normal form is computed.

AlgebraicSolvers.solve — Function

Xi, ms = solve(P, Mth; verbose = false, diag=:schur, refine=true)

P polynomial system

Mth class specifying the solver.

Solve the system of polynomials P=[p1, ..., pn], using Sylvester matrix res_matrix(P,Mth).

Output:

Xi the solutions as a matrix of points, one per column

ms the vector of their multiplicities.

Options:

verbose=::Bool (default false) prints information related to the different substeps and their timing.

diag=::Symbol (default :schur, other possibilities :diag ) specify the method for joint diagonalization of the operators of multiplication.

refine=::Bool (default true) applies Newton iterations to refine the roots.

Example

using AlgebraicSolvers, DynamicPolynomials
X = @polyvar x y
P = [2-x*y+x^2,y^2+x-2]
Xi, ms = solve(P, Macaulay())

Resultant solvers

AlgebraicSolvers.Macaulay — Type

Structure for the construction of Macaulay resultant solvers. It stores

degree :: Function P-> degree ≤ ρ (default: P -> sum(DynamicPolynomials.maxdegree(P[i])-1 for i in 1:length(P)) + 1)

is_homogeneous :: Function P-> boolean testing if the system is homogeneous or not (default: P -> !any(AlgebraicSolvers.is_not_homogeneous, P))

The default value for ρ is ∑ deg(pi) - n + 1.

This specifies

the resultant matrix construction associated to Macaulay construction res_matrix(P,Macaulay())

the TNF which is the transpose of the kernel of the resultant matrix tnf(P,Macaulay())

the solver which is use Macaulay TNF to solve the system solve(P,Macaulay())

It can be used as an argument of solve, tnf, quot_basis.

Constructors

     Macaulay()

The degree function is P -> sum(DynamicPolynomials.maxdegree(P[i])-1 for i in 1:length(P)) + 1

     Macaulay(rho::Int64)

The degree function is P -> rho

Example

using AlgebraicSolvers, DynamicPolynomials
X = @polyvar x y
P = [2-x*y-x^2, y^2+x-2, x^2-y^2+x-1]
Xi, ms = solve(P, Macaulay(3))

This gives the unique solution Xi:

2×1 Matrix{ComplexF64}:
 1.0 + 0.0im
 1.0 + 0.0im

AlgebraicSolvers.Toric — Type

Structure for the construction of Toric resultant solvers. It store

supports :: Function P-> supports of the polynomials P (default: P->AlgebraicSolvers.support.(P) )

This specifies

the resultant matrix associated to the Toric construction res_matrix(P, Toric())

the TNF which is the transpose of the kernel of the resultant matrix tnf(P,Toric())

the solver which is using the Toric TNF to solve the system solve(P,Toric())

Groebner basis solvers

AlgebraicSolvers.GbSolver — Type

Structure which defines

grobner_basis :: Function (P) -> grobner basis of P with variables X

reduce :: Function (p, G) -> normal form of the polynomial p modulo G

quot_basis :: Function P -> basis of the quotient by the ideal(P)

It can be used as an argument of solve, tnf, quot_basis.

Constructors:

    GB = GbSolver(gbasis, normform, qbasis)

Construct a structure of type GbSolver where

gbasis is the function which computes the GbSolver basis

normform is the function to reduce modulo a GbSolver basis

qbasis is the function which provides a basis of the quotient algebra.

    GB = GbSolver()

Construct a GbSolver. If a Groebner engine is available, from a package such as Groebner, it uses it otherwise it returns nothing.

Example:

using AlgebraicSolvers, DynamicPolynomials, Groebner

GB = GbSolver(
        P->Groebner.groebner(P, ordering = Groebner.DegRevLex(variables(P))),
        Groebner.normalform,
        Groebner.quotient_basis
    )

X = @polyvar x y
P = [x-y+x^2,x-y+y^2]
Xi, ms = solve(P, GB)