Solvers

Several types of solvers are available in the packages. They differ in the way the truncated normal form is computed.

AlgebraicSolvers.solve — Function

Xi, ms = solve(P, Mth; verbose = false)

P polynomial system

Mth class specifying the solver.

Solve the system of polynomials P=[p1, ..., pn], using Sylvester matrix res_matrix(P,Mth). It outputs the solutions Xi as a matrix of points, one per column, and the vector of their multiplicities ms.

Xi, ms = solve(P, Macaulay())

P polynomial system

Solve the system P=[p1, ..., pn], building Sylvester matrix of all monomial multiples mi*pi in degree ≤ ρ. It outputs the solutions Xi as a matrix of points, one per column, and the vector of their multiplicities ms.

The default value for ρ is ∑ deg(pi) - n + 1.

Example

using AlgebraicSolvers, DynamicPolynomials

X = @polyvar x y

P = [2-x*y+x^2,y^2+x-2]

Xi = solve(P, Macaulay())

Xi, ms, G, B = solve(P::AbstractVector, Mth::Grobner; verbose = false)

Solve the system of polynomials P. It outputs:

Xi the complex solution points, one per column of Xi

ms the vector of their multiplicities

G the computed Grobner basis

B the basis of the quotient by the ideal of the equations

If verbose = true, the timing of the different steps is printed.

Example:

using AlgebraicSolvers, DynamicPolynomials, Groebner

GB=Grobner(
    Groebner.DegRevLex,
    Groebner.groebner,
    Groebner.normalform,
    Groebner.quotient_basis
)

X = @polyvar x y 

P = [-2+y-y^2+x^2*y, 1-3x+y+x*y^2]

Xi, ms, G, B = AlgebraicSolvers.solve(P, GB; verbose = true)

using AbstractAlgebra

R ,(x,y) =  polynomial_ring(AbstractAlgebra.QQ, [:x,:y])

P = [x^2*y-y^2, x*y^2-3]

Xi, ms, G, B = AlgebraicSolvers.solve(P, GB)

Solvers using resultant constructions

They corresponds to the following type

AlgebraicSolvers.Macaulay — Type

Structure for the construction of Macaulay resultant solvers. It stores

degree :: Function P-> degree of regularity +1 (default: P -> sum(DynamicPolynomials.maxdegree(P[i])-1 for i in 1:length(P)) + 1)

is_homogeneous :: Function P-> boolean testing if the system is homogeneous or not (default: P -> !any(AlgebraicSolvers.is_not_homogeneous, P))

AlgebraicSolvers.Toric — Type

Structure for the construction of Toric resultant solvers. It store

supports :: Function P-> supports of the polynomials P (default: P->AlgebraicSolvers.support.(P) )

Solvers using Groebner basis computation

AlgebraicSolvers.Grobner — Type

Structure which defines

grobner_basis :: Function (P,X) -> grobner basis of P with variables X

reduce :: Function (p, G) -> normal form of the polynomial p modulo G

quotient_basis :: Function P -> basis of the quotient by the ideal(P)