Artinian algebras

Truncated Normal Forms, which kernel generates the ideal, are available:

AlgebraicSolvers.tnf — Function

N, L, B = tnf(P, Mth)

Compute the Truncated Normal Form N of P=[p1, ..., pn], using res_matrix(P,Mth).

The list L is the list of monomials indexing the colmuns of N.

    B = tnf(P, Mth::Grobner)

Computes the Truncated Normal Form on B^+ where B is the basis of the quotient by the ideal (P), formed by the monomials which are not in the inital of (P).

The operators of multiplications (by the variables or any polynomial) can be computed with the following function:

AlgebraicSolvers.mult_matrices — Function

    M = mult_matrices(P, DynamicPolynomials.variables(P), Mth)

Computes the vector of multiplication matrices M=[M1, M2, ...] by the variables in a basis B of the quotient by the ideal (P), using res_matrix(P,Mth).

AlgebraicSolvers.mult_matrix — Function

M = mult_matrix(p, G::AbstractVector, Idx::Dict, M::Grobner)

Compute the matrix of multication by p modulo g in the basis associated to the basis dictionary Idx. It is assumed that g is a Groebner basis and that the quotient is finite dimensional.

M = mult_matrix(p, G::AbstractVector, B::AbstractVector)

Compute the matrix of multication by p modulo G in the basis B. It is assumed that G is a Groebner basis and that the quotient is finite dimensional.

A basis of the quotient by the ideal (P) can be also computed:

AlgebraicSolvers.quot_basis — Function

    B = quot_basis(P, Mth)

Computes a basis of the quotient by the ideal P, using res_matrix(P,Mth)

    B = quot_basis(P, Mth::Grobner)

Computes the basis B of the quotient by the ideal (P), formed by the monomials which are not in the inital of (P).